Create A-DID-for-every-thing---Agile-Driving-Data-Chain
##Submission from Spherity GmbH for RWoT 8

The paper describes an implementation of *DADs* and *verifiable Data Chains* for data provenance in a digital mobility value chain. This work is building on top of the DID for Every Thing work delivered at RWoT 7 and previous work done by Samuel Smith. In addition we are sharing our work on key management for corporate wallets and vaults.