Create Journalistic use-cases for SSI:  signatures, verified claims, and canonical-text registries
The goal of this paper is to compare notes with the SSI technical community and hopefully to model some possible solutions for middleware systems and/or SSI-powered widgets for CMS systems.  Some of these could, we hope, be built soon on current platforms, to start benefiting journalists, factcheckers, and reading publics as soon as possible.  Others, at industry-wide or semi-global scale, could not be built as quickly, but planning them in dialogue with the broader community is also a priority for us.