Create Implementing threshold schemes
A couple of weeks ago, Christopher Allen requested me to "summarize
[my] post https://dsprenkels.com/sss-34c3.html" and submit it to
the [RWOT8 repo][rwot8].
I have not exactly summarized my post. I have left out quite a bit.
However, I have tried to include some background on the
alternatives that exist to Shamir secret sharing, that actually
provide some integrity protection. It would be a shame to restrict
our view to *only* SSS.

Abstract:

Shamir secret sharing is a method to split secrets into shares, and
to later recombine them. However, it does not feature integrity
protection of the secret. This article elaborates on Feldman VSS
and Pederson VSS, which *do* protect the message integrity.
Furthermore, we show how hashing the shares also protects the
message integrity, but is vulnerable to a cheating dealer.